requirements.txt
requests
user-agents
django-user-agents
django-bootstrap-pagination
django-widget-tweaks
